Cleanup code (#38)
* Add blinky main

* First part of cleanup

* Add tempsensor readout timeout and make safetycheck compile

* Add debug messages, remove unnecessary comments

* Fix warnings regarding narrowing conversion

* Add more debugging and cleanup

* Add pwm enable func

* Init PWM buzzer and enable PWM seperatley

* Implement alarm transmits and cutoff pwm and buffer

* Make debug

* Do stuff

* Send correct errors and warnings

---------

+++ ControllerFirmware/include/drivers/current_sensor.hpp
@@ -78,7 +78,7 @@ class INA219 {
 
     private:
     void write_reg(uint8_t reg, uint16_t data) noexcept {
-        _bus->write<3>(_i2c_addr, {(uint8_t)reg, (data >> 8) & 0xFF, data & 0xFF}, true);
+        _bus->write<3>(_i2c_addr, {(uint8_t)reg, static_cast<uint8_t>((data >> 8) & 0xFF), static_cast<uint8_t>(data & 0xFF)}, true);
     }
 
     uint16_t read_reg(uint8_t reg) noexcept {
